+++ stories/Page/ExampleWithAppStructure.js
