+++ stories/Page/PageExampleStories.js
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react';
 import { Category } from '../storiesHierarchy';
 
 import ExampleStandAlone from './ExampleStandAlone';
-import ExampleStandard from './ExampleStandard';
+import ExampleWithAppStructure from './ExampleWithAppStructure';
 import ExampleStretchContent from './ExampleStretchContent';
 import ExampleStretchTable from './ExampleStretchTable';
 
@@ -13,7 +13,7 @@ const PageExampleStories = storiesOf(
 );
 
 PageExampleStories.add('1. StandAlone', () => <ExampleStandAlone />);
-PageExampleStories.add('2. Standard', () => <ExampleStandard />);
+PageExampleStories.add('2. Standard', () => <ExampleWithAppStructure />);
 PageExampleStories.add('3. Stretched Content', () => <ExampleStretchContent />);
 PageExampleStories.add('4. Stretched Table In Page', () => (
   <ExampleStretchTable />
